+++ src/test/java/com/tallerwebi/TDD/ServicioClienteTest.java
@@ -26,7 +26,7 @@ public class ServicioClienteTest {
 
         servicio.registrarCliente (clienteNuevo);
 
-        int idCliente = clienteNuevo.getId();
+        Long idCliente = clienteNuevo.getId();
         int id = 1;
         assertThat(idCliente, is(id));
     }
@@ -87,7 +87,7 @@ public class ServicioClienteTest {
 
         servicio.registrarCliente(clienteNuevo1);
 
-        // 🚀 acá validamos que se lance la excepción
+        //acá validamos que se lance la excepción
         assertThrows(IllegalArgumentException.class, () -> {
             servicio.registrarCliente(clienteNuevo2);
         });
@@ -95,7 +95,7 @@ public class ServicioClienteTest {
 
     @Test
     public void dadoUnClienteConTurnosRegistradosPuedoCancelarUno() {
-        Cliente cliente = new Cliente(202, "Joauquin", "Diaz", "joaquindiazantunez02@gmail.com","1131522182");
+        Cliente cliente = new Cliente(202L, "Joauquin", "Diaz", "joaquindiazantunez02@gmail.com","1131522182");
         Turno t1 = new Turno(Especialidad.CONTROL, Practica.CONTROL_1, 1,"2025-05-14","10:00");
         Turno t2 = new Turno(Especialidad.ESTUDIO, Practica.ESTUDIO_2, 1,"2025-05-15","10:30");
 
@@ -107,7 +107,7 @@ public class ServicioClienteTest {
 
     @Test
     public void dadoUnClienteSinTurnosRegistradosSiIntentoCancelarUnTurnoObtengoUnaExceptionTuroNoCancelado() {
-        Cliente cliente = new Cliente(202, "Joauquin", "Diaz", "joaquindiazantunez02@gmail.com","1131522182");
+        Cliente cliente = new Cliente(202L, "Joauquin", "Diaz", "joaquindiazantunez02@gmail.com","1131522182");
         assertThrows(IllegalArgumentException.class, ()->{
             cliente.cancelarTurno(2);
         });
@@ -115,7 +115,7 @@ public class ServicioClienteTest {
 
     @Test
     public void dadoUnClienteConTurnosSiBuscoUnoRegistradoPorIdLoObtengo() {
-        Cliente cliente = new Cliente(202, "Joauquin", "Diaz", "joaquindiazantunez02@gmail.com","1131522182");
+        Cliente cliente = new Cliente(202L, "Joauquin", "Diaz", "joaquindiazantunez02@gmail.com","1131522182");
         Turno t1 = new Turno(Especialidad.CONTROL, Practica.CONTROL_1, 1,"2025-05-14","10:00");
         Turno t2 = new Turno(Especialidad.ESTUDIO, Practica.ESTUDIO_2, 1,"2025-05-15","10:30");
 
@@ -128,7 +128,7 @@ public class ServicioClienteTest {
 
     @Test
     public void dadoUnClienteSinTurnosRegistradosSiIntentoBuscarUnoPorIdObtengoUnIllegalArgumentException() {
-        Cliente cliente = new Cliente(202, "Joauquin", "Diaz", "joaquindiazantunez02@gmail.com","1131522182");
+        Cliente cliente = new Cliente(202L, "Joauquin", "Diaz", "joaquindiazantunez02@gmail.com","1131522182");
         assertThrows(IllegalArgumentException.class, ()->{
             cliente.getTurnoPorId(1); 
         });
